+++ Assets/Scripts/Board/Tile/TileBoardRandom.cs
@@ -3,13 +3,13 @@ using UnityEngine.Assertions;
 
 public class TileBoardRandom : TileBoard
 {
-    public override void PlayToken(in byte playerId, in byte column)
+    public override void PlayToken(in int playerId, in int column)
     {
         Assert.IsTrue(IsValidPlayerId(playerId), "Invalid player id playing a token");
 
-        byte newCol = (byte)(Mathf.Abs(column + (Random.value*6 - 3)) % width);
+        int newCol = (int)(Mathf.Abs(column + (Random.value*6 - 3)) % width);
 
-        byte nextAvailableRow = CalculateNextAvailableRow(newCol);
+        int nextAvailableRow = CalculateNextAvailableRow(newCol);
         if (nextAvailableRow == height)
             return;
