+++ Assets/Scripts/WinEvaluator/WinEvaluatorConfig.cs
@@ -1,9 +1,9 @@
 public readonly struct WinEvaluatorConfig
 {
     public readonly string name;
-    public readonly byte tokenCountToWin;
+    public readonly int tokenCountToWin;
 
-    public WinEvaluatorConfig(in string name, in byte tokenCountToWin)
+    public WinEvaluatorConfig(in string name, in int tokenCountToWin)
     {
         this.name = name;
         this.tokenCountToWin = tokenCountToWin;
